+++ common/src/main/java/dev/xylonity/bonsai/ghosts/Ghosts.java
@@ -1,9 +1,8 @@
 package dev.xylonity.bonsai.ghosts;
 
 import dev.xylonity.bonsai.ghosts.platform.GhostsPlatform;
-import dev.xylonity.bonsai.ghosts.registry.GhostsEntities;
-import dev.xylonity.bonsai.ghosts.registry.GhostsItems;
-import dev.xylonity.bonsai.ghosts.registry.GhostsSounds;
+import dev.xylonity.bonsai.ghosts.registry.*;
+import net.minecraft.resources.ResourceLocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,9 +16,17 @@ public class Ghosts {
     public static final GhostsPlatform PLATFORM = ServiceLoader.load(GhostsPlatform.class).findFirst().orElseThrow();
 
     public static void init() {
+        GhostsItems.init();
+        GhostsBlocks.init();
+        GhostsCreativeTabs.init();
         GhostsEntities.init();
         GhostsSounds.init();
-        GhostsItems.init();
+        GhostsTreeDecoratorTypes.init();
+        GhostsFoliagePlacers.init();
+    }
+
+    public static ResourceLocation of(String path) {
+        return new ResourceLocation(MOD_ID, path);
     }
 
 }