+++ common/src/main/java/dev/xylonity/bonsai/ghosts/configurations/tree/HauntedTreeGrower.java
@@ -1,0 +1,18 @@
+package dev.xylonity.bonsai.ghosts.configurations.tree;
+
+import dev.xylonity.bonsai.ghosts.registry.GhostsConfiguredFeatures;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.block.grower.AbstractTreeGrower;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+
+import javax.annotation.Nullable;
+
+public class HauntedTreeGrower extends AbstractTreeGrower {
+
+    @Override
+    protected @Nullable ResourceKey<ConfiguredFeature<?, ?>> getConfiguredFeature(RandomSource randomSource, boolean b) {
+        return GhostsConfiguredFeatures.HAUNTED_TREE;
+    }
+
+}