+++ fabric/src/main/java/dev/xylonity/bonsai/ghosts/common/event/GhostsServerEvents.java
@@ -15,12 +15,6 @@ public class GhostsServerEvents {
         FabricDefaultAttributeRegistry.register(GhostsEntities.GHOST.get(), GhostEntity.setAttributes());
         FabricDefaultAttributeRegistry.register(GhostsEntities.SMALL_GHOST.get(), SmallGhostEntity.setAttributes());
         FabricDefaultAttributeRegistry.register(GhostsEntities.KODAMA.get(), KodamaEntity.setAttributes());
-
-        ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.SPAWN_EGGS).register(entries -> {
-            entries.accept(GhostsItems.GHOST_SPAWN_EGG.get());
-            entries.accept(GhostsItems.SMALL_GHOST_SPAWN_EGG.get());
-            entries.accept(GhostsItems.KODAMA_SPAWN_EGG.get());
-        });
     }
 
 }