+++ forge/src/main/java/dev/xylonity/bonsai/ghosts/common/datagen/GhostsWorldgenProvider.java
@@ -1,0 +1,23 @@
+package dev.xylonity.bonsai.ghosts.common.datagen;
+
+import dev.xylonity.bonsai.ghosts.Ghosts;
+import dev.xylonity.bonsai.ghosts.registry.GhostsConfiguredFeatures;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.PackOutput;
+import net.minecraftforge.common.data.DatapackBuiltinEntriesProvider;
+
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+public class GhostsWorldgenProvider extends DatapackBuiltinEntriesProvider {
+
+    public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+            .add(Registries.CONFIGURED_FEATURE, GhostsConfiguredFeatures::register);
+
+    public GhostsWorldgenProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
+        super(output, registries, BUILDER, Set.of(Ghosts.MOD_ID));
+    }
+
+}