+++ forge/src/main/java/dev/xylonity/bonsai/ghosts/platform/GhostsPlatformForge.java
@@ -1,12 +1,22 @@
 package dev.xylonity.bonsai.ghosts.platform;
 
+import com.mojang.serialization.Codec;
 import dev.xylonity.bonsai.ghosts.Ghosts;
 import dev.xylonity.bonsai.ghosts.GhostsForge;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.*;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.levelgen.feature.foliageplacers.FoliagePlacer;
+import net.minecraft.world.level.levelgen.feature.foliageplacers.FoliagePlacerType;
+import net.minecraft.world.level.levelgen.feature.treedecorators.TreeDecorator;
+import net.minecraft.world.level.levelgen.feature.treedecorators.TreeDecoratorType;
+import net.minecraft.world.level.levelgen.feature.trunkplacers.TrunkPlacerType;
 import net.minecraftforge.common.ForgeSpawnEggItem;
+import net.minecraftforge.registries.RegistryObject;
 
 import javax.annotation.Nullable;
 import java.util.List;
@@ -19,6 +29,20 @@ public class GhostsPlatformForge implements GhostsPlatform {
     @Override
     public <X extends Item> Supplier<X> registerItem(String id, Supplier<X> item) {
         return GhostsForge.ITEMS.register(id, item);
+    }
+
+    @Override
+    public <X extends Block> Supplier<X> registerBlock(String id, Supplier<X> block) {
+        Supplier<X> blockSupplier = GhostsForge.BLOCKS.register(id, block);
+
+        registerItem(id, () -> new BlockItem(blockSupplier.get(), new Item.Properties()));
+
+        return blockSupplier;
+    }
+
+    @Override
+    public <X extends CreativeModeTab> Supplier<X> registerCreativeTab(String id, Supplier<X> creativeModeTab) {
+        return GhostsForge.CREATIVE_TAB.register(id, creativeModeTab);
     }
 
     @Override
@@ -44,6 +68,26 @@ public class GhostsPlatformForge implements GhostsPlatform {
     @Override
     public <X extends SoundEvent> Supplier<X> registerSound(String id, Supplier<X> sound) {
         return GhostsForge.SOUNDS.register(id, sound);
+    }
+
+    @Override
+    public <X extends TrunkPlacerType<?>> Supplier<X> registerTrunkPlacer(String id, Supplier<X> item) {
+        return null;
+    }
+
+    @Override
+    public <U extends TreeDecorator> Supplier<TreeDecoratorType<U>> registerTreeDecorator(String id, Codec<U> codec) {
+        return GhostsForge.TREE_DECORATOR_TYPES.register(id, () -> new TreeDecoratorType<>(codec));
+    }
+
+    @Override
+    public <U extends FoliagePlacer> Supplier<FoliagePlacerType<U>> registerFoliagePlacer(String id, Codec<U> codec) {
+        return GhostsForge.FOLIAGE_TYPES.register(id, () -> new FoliagePlacerType<>(codec));
+    }
+
+    @Override
+    public CreativeModeTab.Builder creativeTabBuilder() {
+        return CreativeModeTab.builder();
     }
 
 }