+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - Added calibrated eye block, that outputs a redstone signal if there is an entity in front of it
 - Fixed z fighting on the small ghost plant texture
 - Made regular ghosts slightly more transparent
+- Now small ghosts do not phase through blocks
 - Fixed small ghost red hurt overlay not showing up on its body